CM-505 Refactoring. Fix CreateStep4
