+++ src/Graphql/index.ts
@@ -1,5 +1,6 @@
 import {gql} from '@apollo/client';
 import * as Apollo from '@apollo/client';
+import {Common, CommonContributionType} from './Common';
 
 export type Maybe<T> = T | null;
 export type Exact<T extends {[key: string]: unknown}> = {[K in keyof T]: T[K]};
@@ -133,20 +134,6 @@ export type Proposal = {
   proposer: User;
 };
 
-// countdownPeriod: number;
-// createdAt: Time;
-// description: { description: string; links: Rules[]; images: File[]; files: File[]; title: string };
-// fundingRequest: { funded: boolean; amount: number };
-// fundingState: string;
-// id: string;
-// moderation: Moderation;
-// paymentState: string;
-// proposerId: string;
-// quietEndingPeriod: 10800;
-// state: string;
-// type: string;
-// updatedAt: Time;
-
 export type ProposalDescription = {
   __typename?: 'ProposalDescription';
   title?: Maybe<Scalars['String']>;
@@ -224,29 +211,6 @@ export type User = {
   subscriptions?: Maybe<Array<Maybe<Subscription>>>;
 };
 
-export enum CommonContributionType {
-  OneTime = 'OneTime',
-  Monthly = 'Monthly',
-}
-
-export type CommonMetadata = {
-  __typename?: 'CommonMetadata';
-  byline: Scalars['String'];
-  description: Scalars['String'];
-  founderId: Scalars['String'];
-  minFeeToJoin: Scalars['Int'];
-  contributionType: CommonContributionType;
-  action: Scalars['String'];
-};
-
-export type CommonMember = {
-  __typename?: 'CommonMember';
-  /** The user ID of the member */
-  userId: Scalars['ID'];
-  /** The date, at witch the member joined the common */
-  joinedAt?: Maybe<Scalars['Date']>;
-};
-
 export type Link = {
   title: Scalars['String'];
   url: Scalars['String'];
@@ -255,34 +219,6 @@ export type Link = {
 export type Rule = {
   title: Scalars['String'];
   description: Scalars['String'];
-};
-
-/** The common type */
-export type Common = {
-  __typename?: 'Common';
-  /** The unique identifier of the common */
-  id: Scalars['ID'];
-  /** The date, at which the common was created */
-  createdAt?: Maybe<Scalars['Date']>;
-  /** The date, at which the common was last updated */
-  updatedAt?: Maybe<Scalars['Date']>;
-  /** The display name of the common */
-  name: Scalars['String'];
-  /** The currently available funds of the common */
-  balance: Scalars['Int'];
-  /** The total amount of money, raised by the common */
-  raised: Scalars['Int'];
-  byline: Scalars['String'];
-  description: Scalars['String'];
-  fundingMinimumAmount: Scalars['Int'];
-  links: Array<Link>;
-  fundingType: CommonContributionType;
-  members: Array<CommonMember>;
-  proposals?: Maybe<Array<Maybe<Proposal>>>;
-  openJoinRequests: Scalars['Int'];
-  openFundingRequests: Scalars['Int'];
-  image: Scalars['String'];
-  register: Scalars['String'];
 };
 
 export type Mutation = {
@@ -295,157 +231,9 @@ export type Mutation = {
   createDiscussion: Discussion;
 };
 
-export type GetCommonsDataQuery = {__typename?: 'Query'} & {
-  commons?: Array<
-    {__typename?: 'Common'} & Common & {
-        members?: Maybe<
-          Array<
-            Maybe<{__typename?: 'CommonMember'} & Pick<CommonMember, 'userId'>>
-          >
-        >;
-      }
-  >;
-};
-
-export const GetCommonsDataDocument = gql`
-  query getCommonsHomescreenData(
-    $paginate: PaginateInput! = {take: 10, skip: 0}
-  ) {
-    commons(paginate: $paginate) {
-      id
-      name
-      createdAt
-      updatedAt
-      whitelisted
-      members {
-        userId
-      }
-      proposals {
-        id
-      }
-    }
-  }
-`;
-
-export const GetCommonProposals = gql`
-  query getCommonProposals(
-    $where: ProposalWhereInput
-    $paginate: PaginateInput! = {take: 10, skip: 0}
-  ) {
-    proposals(where: $where, paginate: $paginate) {
-      id
-      state
-      createdAt
-      updatedAt
-      links
-      files
-      images
-      votesFor
-      votesAgainst
-      title
-      description
-      discussions {
-        id
-        topic
-        description
-        latestMessage
-        type
-        userId
-        owner {
-          id
-          displayName
-          lastName
-          firstName
-        }
-        messages {
-          message
-          type
-          flag
-          reports {
-            status
-            message {
-              message
-              type
-              flag
-            }
-            for
-            note
-            reporterId
-          }
-        }
-      }
-      funding {
-        id
-        fundingState
-        amount
-      }
-      join {
-        id
-        funding
-        fundingType
-        paymentState
-      }
-    }
-  }
-`;
-
 export type DiscussionWhereInput = {
   commonId: Scalars['ID'];
 };
-
-export type GetCommonDiscussionsQueryVariables = Exact<{
-  where: DiscussionWhereInput;
-}>;
-
-export const GetCommonDiscussions = gql`
-  query getCommonDiscussions(
-    $where: DiscussionWhereInput
-    $paginate: PaginateInput! = {take: 10, skip: 0}
-  ) {
-    discussions(where: $where, paginate: $paginate) {
-      id
-      createdAt
-      messages {
-        id
-        message
-        createdAt
-        updatedAt
-        type
-        flag
-        owner {
-          displayName
-          firstName
-          lastName
-          photo
-        }
-      }
-      title: topic
-      description
-      userId
-      owner {
-        firstName
-        lastName
-        displayName
-        photo
-      }
-    }
-  }
-`;
-export type GetCommonDiscussionsQuery = {
-  discussions?: Maybe<Array<Discussion>>;
-};
-
-export function useGetCommonDiscussions(
-  baseOptions: Apollo.QueryHookOptions<
-    GetCommonDiscussionsQuery,
-    GetCommonDiscussionsQueryVariables
-  >,
-) {
-  return Apollo.useQuery<
-    GetCommonDiscussionsQuery,
-    GetCommonDiscussionsQueryVariables
-  >(GetCommonDiscussions, baseOptions);
-}
 
 export const GetDiscussionById = gql`
   query getDiscussionById($id: ID!) {
@@ -479,10 +267,6 @@ export const GetDiscussionById = gql`
   }
 `;
 
-export type GetCommonProposalsQuery = {
-  proposals?: Maybe<Array<Proposal>>;
-};
-
 export const GetUserPermissionsDocument = gql`
   query getUserPermissions($userId: ID!) {
     user(id: $userId) {
@@ -491,69 +275,9 @@ export const GetUserPermissionsDocument = gql`
   }
 `;
 
-export type GetCommonByIdQuery = {
-  common?: Maybe<Common>;
-};
-
-export type CommonWhereUniqueInput = {
-  id: Scalars['ID'];
-};
-
-export type GetCommonByIdQueryVariables = Exact<{
-  where: CommonWhereUniqueInput;
-}>;
-
-export const GetCommonByIdDocument = gql`
-  query getCommon($where: CommonWhereUniqueInput!) {
-    common(where: $where) {
-      id
-      name
-      image
-      balance
-      raised
-      members {
-        userId
-        joinedAt: createdAt
-        roles
-        user {
-          id
-        }
-      }
-      rules
-      links
-      register: whitelisted
-      action
-      byline
-      description
-      contributionType: fundingType
-      minFeeToJoin: fundingMinimumAmount
-    }
-  }
-`;
-
 export type GetUserPermissionsQuery = {__typename?: 'Query'} & {
   user?: Maybe<{__typename?: 'User'} & Pick<User, 'permissions'>>;
 };
-
-export type GetCommonDataQueryVariables = Exact<{
-  paginate?: Pagination;
-}>;
-
-export type GetCommonDataQuery = {
-  commons?: Maybe<Array<Common>>;
-};
-
-export function useGetCommonDataQuery(
-  baseOptions: Apollo.QueryHookOptions<
-    GetCommonDataQuery,
-    GetCommonDataQueryVariables
-  >,
-) {
-  return Apollo.useQuery<GetCommonDataQuery, GetCommonDataQueryVariables>(
-    GetCommonsDataDocument,
-    baseOptions,
-  );
-}
 
 export type GetUserPermissionsQueryVariables = Exact<{
   userId: Scalars['ID'];
@@ -563,17 +287,6 @@ export type Pagination = {
   skip: Scalars['Int'];
   take: Scalars['Int'];
 };
-
-export type GetCommonProposalsQueryVariables = Exact<{
-  where: {
-    type?: ProposalType;
-    state?: ProposalState;
-    commonId?: Scalars['ID'];
-    commonMemberId?: Scalars['ID'];
-    userId?: Scalars['ID'];
-  };
-  paginate?: Pagination;
-}>;
 
 export function useGetUserPermissionsQuery(
   baseOptions: Apollo.QueryHookOptions<
@@ -585,30 +298,6 @@ export function useGetUserPermissionsQuery(
     GetUserPermissionsQuery,
     GetUserPermissionsQueryVariables
   >(GetUserPermissionsDocument, baseOptions);
-}
-
-export function useGetCommonProposals(
-  baseOptions: Apollo.QueryHookOptions<
-    GetCommonProposalsQuery,
-    GetCommonProposalsQueryVariables
-  >,
-) {
-  return Apollo.useQuery<
-    GetCommonProposalsQuery,
-    GetCommonProposalsQueryVariables
-  >(GetCommonProposals, baseOptions);
-}
-
-export function useGetCommonById(
-  baseOptions: Apollo.QueryHookOptions<
-    GetCommonByIdQuery,
-    GetCommonByIdQueryVariables
-  >,
-) {
-  return Apollo.useQuery<GetCommonByIdQuery, GetCommonByIdQueryVariables>(
-    GetCommonById,
-    baseOptions,
-  );
 }
 
 export const LoadUserContextDocument = gql`
@@ -735,50 +424,6 @@ export function useUpdateUserMutation(
   );
 }
 
-export type CreateCommonInput = {
-  name: Scalars['String'];
-  fundingMinimumAmount: Scalars['Int'];
-  fundingType: Maybe<CommonContributionType>;
-  image: Scalars['String'];
-  description: Scalars['String'];
-  action: Scalars['String'];
-  byline: Scalars['String'];
-  links: Array<Link>;
-  rules: Array<Rule>;
-};
-
-export type CreateCommonMutation = {__typename?: 'Mutation'} & Pick<
-  Mutation,
-  'createCommon'
->;
-
-export type CreateCommonMutationVariables = Exact<{
-  common: CreateCommonInput;
-}>;
-
-export const CreateCommonDocument = gql`
-  mutation CreateCommon($common: CreateCommonInput!) {
-    createCommon(input: $common) {
-      id
-      name
-      links
-      rules
-    }
-  }
-`;
-
-export function useCreateCommonMutation(
-  baseOptions?: Apollo.MutationHookOptions<
-    CreateCommonMutation,
-    CreateCommonMutationVariables
-  >,
-) {
-  return Apollo.useMutation<
-    CreateCommonMutation,
-    CreateCommonMutationVariables
-  >(CreateCommonDocument, baseOptions);
-}
-
 export type CreateDiscussionInput = {
   topic: Scalars['String'];
   description: Scalars['String'];
@@ -815,46 +460,6 @@ export function useCreateDiscussionMutation(
   >(CreateDiscussionDocument, baseOptions);
 }
 
-export type UpdateCommonInfoInput = {
-  commonId: Scalars['String'];
-  name?: Scalars['String'];
-  image?: Scalars['String'];
-  action?: Scalars['String'];
-  byline?: Scalars['String'];
-  description?: Scalars['String'];
-  links: Array<Link>;
-  rules: Array<Rule>;
-};
-
-export type UpdateCommonInfoMutation = {__typename?: 'Mutation'} & Pick<
-  Mutation,
-  'updateCommon'
->;
-
-export type UpdateCommonInfoMutationVariables = Exact<{
-  common: UpdateCommonInfoInput;
-}>;
-
-export const UpdateCommonInfoDocument = gql`
-  mutation UpdateCommonInfo($common: UpdateCommonInfoInput!) {
-    updateCommon(input: $common) {
-      id
-    }
-  }
-`;
-
-export function useUpdateCommonInfoMutation(
-  baseOptions?: Apollo.MutationHookOptions<
-    UpdateCommonInfoMutation,
-    UpdateCommonInfoMutationVariables
-  >,
-) {
-  return Apollo.useMutation<
-    UpdateCommonInfoMutation,
-    UpdateCommonInfoMutationVariables
-  >(UpdateCommonInfoDocument, baseOptions);
-}
-
 export type UserWhereUniqueInput = {
   userId: Scalars['ID'];
 };
@@ -883,104 +488,6 @@ export const GetUserInfoDocument = gql`
         displayName
         photo
       }
-    }
-  }
-`;
-
-export const GetUserPendingCommonsDocument = gql`
-  query PendingCommons {
-    user {
-      proposals(where: {state: ${ProposalState.Countdown}, type: ${ProposalType.JoinRequest}}) {
-        common {
-          id
-          name
-          image
-          balance
-          raised
-          members {
-            userId
-            joinedAt: createdAt
-            roles
-            user {
-              id
-            }
-          }
-          rules
-          links
-          register: whitelisted
-          action
-          byline
-          description
-          contributionType: fundingType
-          minFeeToJoin: fundingMinimumAmount
-        }
-      }
-    }
-  }
-`;
-
-export const GetUserCommonsDocument = gql`
-  query MyCommons {
-    user {
-      commons {
-        id
-        name
-        image
-        balance
-        raised
-        members {
-          userId
-          joinedAt: createdAt
-          roles
-          user {
-            id
-          }
-        }
-        rules
-        links
-        register: whitelisted
-        action
-        byline
-        description
-        contributionType: fundingType
-        minFeeToJoin: fundingMinimumAmount
-      }
-    }
-  }
-`;
-
-export type CommonsWhereInput = {
-  ids: Array<string>;
-  page: number;
-};
-
-export const GetCommonsDocument = gql`
-  query AllCommons(
-    $where: [String!]
-    $paginate: PaginateInput! = {take: 10, skip: 0}
-  ) {
-    commons(where: {id: {notIn: $where}}, paginate: $paginate) {
-      id
-      name
-      image
-      balance
-      raised
-      members {
-        userId
-        joinedAt: createdAt
-        roles
-        user {
-          id
-        }
-      }
-      rules
-      links
-      register: whitelisted
-      action
-      byline
-      description
-      contributionType: fundingType
-      minFeeToJoin: fundingMinimumAmount
     }
   }
 `;
